requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
plotly>=5.3.0
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0
aiohttp>=3.8.0
python-binance>=1.0.15

# Database
pymongo>=4.6.0
motor>=3.3.0  # Async MongoDB driver

# Data processing and analysis
scipy>=1.7.0
ta>=0.7.0  # Technical analysis indicators

# Visualization
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# Reporting
jinja2>=3.0.0
pdfkit>=1.0.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.16.0

# Development
black>=21.9b0
flake8>=3.9.0
mypy>=0.910

# Performance visualization
kaleido==0.2.1  # For static image export

# Data handling and analysis
python-dotenv==1.0.0
pyyaml==6.0.1

# CLI Tools
click==8.1.7
rich==13.7.0
typer==0.9.0  # Alternative to Click for more complex CLIs

# Reporting
weasyprint==59.0  # For PDF export

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
structlog==23.1.0

# Testing
pytest-cov==4.1.0

# Added from the code block
flask==3.0.2
pymongo==4.6.2
plotly==5.19.0
rich==13.7.1
PyYAML==6.0.1
python-dotenv==1.0.1
Flask-PyMongo==2.3.0
numpy==1.26.4
requests==2.31.0
python-dateutil==2.9.0
Werkzeug==3.0.1
Jinja2==3.1.3
MarkupSafe==2.1.5
itsdangerous==2.1.2
blinker==1.7.0 